add stories and random function to RestaurantCard
