+++ src/components/RestaurantCard.stories.js
@@ -1,0 +1,11 @@
+import RestaurantCard from './RestaurantCard.js';
+
+export default {
+  title: 'components/RestaurantCard',
+  component: RestaurantCard,
+};
+
+const Template = args => <RestaurantCard {...args} />;
+
+export const Default = Template.bind({});
+Default.args = {};
